Prepare hashcat_get_status()
