+++ include/hashcat.h
@@ -6,9 +6,17 @@
 #ifndef _HASHCAT_H
 #define _HASHCAT_H
 
-int  hashcat (hashcat_ctx_t *hashcat_ctx, char *install_folder, char *shared_folder, int argc, char **argv, const int comptime);
+int   hashcat_init               (hashcat_ctx_t *hashcat_ctx, void (*event) (const u32, struct hashcat_ctx *, const void *, const size_t));
+void  hashcat_destroy            (hashcat_ctx_t *hashcat_ctx);
 
-int  hashcat_ctx_init    (hashcat_ctx_t *hashcat_ctx, void (*event) (const u32, struct hashcat_ctx *, const void *, const size_t));
-void hashcat_ctx_destroy (hashcat_ctx_t *hashcat_ctx);
+int   hashcat_session_run        (hashcat_ctx_t *hashcat_ctx, char *install_folder, char *shared_folder, int argc, char **argv, const int comptime);
+int   hashcat_session_pause      (hashcat_ctx_t *hashcat_ctx);
+int   hashcat_session_resume     (hashcat_ctx_t *hashcat_ctx);
+int   hashcat_session_bypass     (hashcat_ctx_t *hashcat_ctx);
+int   hashcat_session_checkpoint (hashcat_ctx_t *hashcat_ctx);
+int   hashcat_session_quit       (hashcat_ctx_t *hashcat_ctx);
+
+char *hashcat_get_log            (hashcat_ctx_t *hashcat_ctx);
+int   hashcat_get_status         (hashcat_ctx_t *hashcat_ctx, hashcat_status_t *hashcat_status);
 
 #endif // _HASHCAT_H
