+++ include/restore.h
@@ -29,8 +29,6 @@ int cycle_restore (hashcat_ctx_t *hashcat_ctx);
 
 void unlink_restore (hashcat_ctx_t *hashcat_ctx);
 
-void stop_at_checkpoint (hashcat_ctx_t *hashcat_ctx);
-
 int restore_ctx_init (hashcat_ctx_t *hashcat_ctx, int argc, char **argv);
 
 void restore_ctx_destroy (hashcat_ctx_t *hashcat_ctx);
